+++ lib/rubocop/cop/lint/assignment_in_condition.rb
@@ -48,6 +48,7 @@ module RuboCop
 
         def skip_children?(asgn_node)
           (asgn_node.send_type? && asgn_node.method_name !~ /=\z/) ||
+            empty_condition?(asgn_node) ||
             (safe_assignment_allowed? && safe_assignment?(asgn_node))
         end